ensure pitest and jdk cannot be instrumented
These checks would only come into play if the user set an over wide
target glob, but were previously broken.
