+++ pitest-entry/src/test/java/com/example/systemtest/FailingTest.java
@@ -1,0 +1,12 @@
+package com.example.systemtest;
+
+import org.pitest.simpletest.TestAnnotationForTesting;
+
+import static org.junit.Assert.assertEquals;
+
+public class FailingTest {
+    @TestAnnotationForTesting
+    public void fail() {
+        assertEquals(1, 2);
+    }
+}
