+++ pitest-entry/src/test/java/com/example/systemtest/NoMutationsTest.java
@@ -1,0 +1,10 @@
+package com.example.systemtest;
+
+import org.pitest.simpletest.TestAnnotationForTesting;
+
+public class NoMutationsTest {
+    @TestAnnotationForTesting
+    public void pass() {
+
+    }
+}
