+++ pitest-entry/src/test/java/com/example/systemtest/OneMutationFullTest.java
@@ -1,0 +1,12 @@
+package com.example.systemtest;
+
+import org.pitest.simpletest.TestAnnotationForTesting;
+
+import static org.junit.Assert.assertEquals;
+
+public class OneMutationFullTest {
+    @TestAnnotationForTesting
+    public void testReturnOne() {
+        assertEquals(1, OneMutationOnly.returnOne());
+    }
+}
