+++ pitest-entry/src/test/java/com/example/systemtest/OneMutationOnly.java
@@ -1,0 +1,7 @@
+package com.example.systemtest;
+
+public class OneMutationOnly {
+    public static int returnOne() {
+        return 1;
+    }
+}
