+++ pitest-entry/src/test/java/com/example/systemtest/OneMutationTest.java
@@ -1,0 +1,5 @@
+package com.example.systemtest;
+
+public class OneMutationTest {
+
+}
